+++ internal/core/account.go
@@ -5,6 +5,6 @@ import "time"
 type Account struct {
 	ID        int
 	Name      string
-	Balance   float64
+	Balance   int64
 	CreatedAt time.Time
 }
